auto-reset package branches to master or next for packages that are not part of next/stable
This code checks that, if one is on the next flavor, packages that
are *not* part of next are *not* using the next branch.

The reason is that it reduces the amount of possible errors in the
source.yml, as a common "branch: $ROCK_FLAVOR" case can exist, while
non-next or non-stable packages will automatically stay on the master
branch.