improve flavor support w.r.t. deprecation and imported package sets
The major change is that packages are "included" in flavors only
for package sets that declare that they are part of the flavor
system. Other package sets are left untouched.

As a side effect to that change, the remove_packages_from_flavor
call allows to explicitly remove a package from a flavor (including
implicit flavors), allowing to deprecate it.