extract the Eigen ruby bindings in a separate package
This will allow releasing them as a gem, thus improving the integration
with both ruby-doc and travis-ci for this package and for other pure-ruby
packages that depend on it (but not on base/types)

Having it installed at the same time than base/types should be
harmless (it will pick one or the other depending on the setup of
RUBYLIB)

The library has had its test suite extended, which uncovered a few
bugs. Documentation has been improved dramatically by documenting
the C++ extension.