handle orogen's --test option
Once upon a time, a few years ago, Matthias implemented namespace
support in orogen. The main goal was to allow to define components
in a 'test' namespace, which would be disabled by default and only
enabled with orogen's `--test` option.

This links the `--test` option with orogen's test facility to use
the capability.