provide an alternate way to pass test configuration to Ruby packages
So far, we have only been providing the raw command-line options
necessary to get JUnit reports for Ruby packages using minitest.
This has been pretty effective as it requires no changes to
the packages themselves.

However, it turns out that I'd like to add the possibility to run
linters, as well as handle more complex situations (like Syskit
or bundles, which do separate runs of multiple test suites). This
does not work with the current setup.

So, *in addition to the existing flags*, pass "semantic" flags
RUBOCOP, JUNIT and REPORT_DIR to allow specific packages to use
them and create their own configurations, more adapted to their
specific needs. This should have zero effect on existing packages
which will simply ignore the flags.