fix QtRuby support in Ubuntu 20.04
The writing was on the wall, but I personally always postponed working
on it. Focal removed Qt4 packages. This is the immediate workaround.

This adds reference to a PPA I just created with the eon Qt4 packages
slightly adapted and rebuilt for focal. In addition, we need a (also
slightly modified) qtbindings version as the Ruby 2.7 ruby headers use
`alignof` which is only defined in C++11 (incompatible with Qt 4.8)